Push matinale du 08/06

+++ src/Controller/SortieController.php
@@ -92,14 +92,12 @@ class SortieController extends AbstractController
                 $etat = $etatRepository->findOneBy(['code' => 'CREA']);
                 $sortie->setEtat($etat);
                 $sortieRepository->add($sortie, true);
-                //return $this->redirectToRoute('home');
             }
 
             if ($form->get('publish')->isClicked()) {
                 $etat = $etatRepository->findOneBy(['code' => 'O']);
                 $sortie->setEtat($etat);
                 $sortieRepository->add($sortie, true);
-                //return $this->redirectToRoute('home');
             }
 
         }
@@ -118,17 +116,16 @@ class SortieController extends AbstractController
     {
         $sortie = $sortieRepository->find($id);
 
-//        $form = $this->createForm(cancelActivityType::class, $sortie);
+        $form = $this->createForm(cancelActivityType::class, $sortie);
 
-//        $form->handleRequest($request);
+        $form->handleRequest($request);
 
-//        if ($form->isSubmitted() && $form->isValid()) {
+        if ($form->isSubmitted() && $form->isValid()) {
 
-//            if ($form->get('delete')->isClicked()) {
+            if ($form->get('delete')->isClicked()) {
                 $sortieRepository->remove($sortie, true);
-                return $this->redirectToRoute('home');
-//            }
-//        }
+           }
+        }
 
         return $this->render('sortie/cancel_sortie.html.twig', [
             'controller_name' => 'SortieController',
